add SpeckleHttpTests
