+++ tests/Speckle.Sdk.Tests.Integration/Api/GraphQL/Resources/ActiveUserResourceTests.cs
@@ -3,7 +3,6 @@ using Speckle.Sdk.Api;
 using Speckle.Sdk.Api.GraphQL.Inputs;
 using Speckle.Sdk.Api.GraphQL.Resources;
 
-
 namespace Speckle.Sdk.Tests.Integration.API.GraphQL.Resources;
 
 public class ActiveUserResourceTests : IAsyncLifetime
