+++ tests/Speckle.Sdk.Tests.Integration/Api/GraphQL/Resources/ProjectInviteResourceExceptionalTests.cs
@@ -4,7 +4,6 @@ using Speckle.Sdk.Api.GraphQL.Inputs;
 using Speckle.Sdk.Api.GraphQL.Models;
 using Speckle.Sdk.Api.GraphQL.Resources;
 
-
 namespace Speckle.Sdk.Tests.Integration.API.GraphQL.Resources;
 
 public class ProjectInviteResourceExceptionalTests : IAsyncLifetime
