+++ tests/Speckle.Sdk.Tests.Unit/Api/GraphQL/Client.cs
@@ -1,6 +1,3 @@
 ﻿namespace Speckle.Sdk.Tests.Unit.Api.GraphQL;
 
-public class Client
-{
-  
-}
+public class Client { }
