+++ tests/Speckle.Sdk.Tests.Unit/Api/Operations/SendReceiveLocal.cs
@@ -7,7 +7,6 @@ using Speckle.Sdk.Models;
 using Speckle.Sdk.Tests.Unit.Host;
 using Speckle.Sdk.Transports;
 
-
 namespace Speckle.Sdk.Tests.Unit.Api.Operations;
 
 public sealed class SendReceiveLocal : IDisposable
