+++ tests/Speckle.Sdk.Tests.Unit/Api/Operations/SerializationTests.cs
@@ -5,7 +5,6 @@ using Speckle.Sdk.Api;
 using Speckle.Sdk.Host;
 using Speckle.Sdk.Models;
 using Speckle.Sdk.Tests.Unit.Host;
-
 using Point = Speckle.Sdk.Tests.Unit.Host.Point;
 
 namespace Speckle.Sdk.Tests.Unit.Api.Operations;
