+++ tests/Speckle.Sdk.Tests.Unit/Assembly.cs
@@ -1,3 +1,1 @@
-﻿
-
-[assembly: CollectionBehavior(DisableTestParallelization = true)]
+﻿[assembly: CollectionBehavior(DisableTestParallelization = true)]
