+++ tests/Speckle.Sdk.Tests.Unit/Models/Extensions/DisplayValueTests.cs
@@ -3,7 +3,6 @@ using Speckle.Sdk.Host;
 using Speckle.Sdk.Models;
 using Speckle.Sdk.Models.Extensions;
 
-
 namespace Speckle.Sdk.Tests.Unit.Models.Extensions;
 
 public class DisplayValueTests
