+++ tests/Speckle.Sdk.Tests.Unit/Models/GraphTraversal/TraversalContextExtensionsTests.cs
@@ -4,7 +4,6 @@ using Speckle.Sdk.Models;
 using Speckle.Sdk.Models.Collections;
 using Speckle.Sdk.Models.GraphTraversal;
 
-
 namespace Speckle.Sdk.Tests.Unit.Models.GraphTraversal;
 
 // Mark test class for xUnit
