+++ tests/Speckle.Sdk.Tests.Unit/Serialisation/BatchTests.cs
@@ -2,7 +2,6 @@ using FluentAssertions;
 using Speckle.Sdk.Dependencies;
 using Speckle.Sdk.Serialisation.V2.Send;
 
-
 namespace Speckle.Sdk.Tests.Unit.Serialisation;
 
 public class BatchTests
