+++ tests/Speckle.Sdk.Tests.Unit/Serialisation/ChunkingTests.cs
@@ -5,7 +5,6 @@ using Speckle.Sdk.Models;
 using Speckle.Sdk.Serialisation;
 using Speckle.Sdk.Transports;
 
-
 namespace Speckle.Sdk.Tests.Unit.Serialisation;
 
 public class ChunkingTests
