+++ tests/Speckle.Sdk.Tests.Unit/Serialisation/JsonIgnoreAttributeTests.cs
@@ -6,7 +6,6 @@ using Speckle.Sdk.Models;
 using Speckle.Sdk.Serialisation;
 using Speckle.Sdk.Transports;
 
-
 namespace Speckle.Sdk.Tests.Unit.Serialisation;
 
 /// <summary>
