+++ tests/Speckle.Sdk.Tests.Unit/Serialisation/SerializerBreakingChanges.cs
@@ -6,7 +6,6 @@ using Speckle.Sdk.Models;
 using Speckle.Sdk.Serialisation;
 using Speckle.Sdk.Tests.Unit.Host;
 
-
 namespace Speckle.Sdk.Tests.Unit.Serialisation;
 
 /// <summary>
