+++ tests/Speckle.Sdk.Tests.Unit/Transports/DiskTransportTests.cs
@@ -2,7 +2,6 @@ using FluentAssertions;
 using Speckle.Sdk.Common;
 using Speckle.Sdk.Transports;
 
-
 namespace Speckle.Sdk.Tests.Unit.Transports;
 
 public sealed class DiskTransportTests : TransportTests, IDisposable
