+++ tests/Speckle.Sdk.Tests.Unit/Transports/TransportTests.cs
@@ -3,7 +3,6 @@ using Speckle.Newtonsoft.Json;
 using Speckle.Sdk.Common;
 using Speckle.Sdk.Transports;
 
-
 namespace Speckle.Sdk.Tests.Unit.Transports;
 
 public abstract class TransportTests
